[FEATURE]: add teansform to the logs table
