+++ logstable/src/model.ts
@@ -11,8 +11,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { LogData, ThresholdOptions } from '@perses-dev/core';
-import { PanelProps, LegendSpecOptions } from '@perses-dev/plugin-system';
+import { LogData, ThresholdOptions, Transform } from '@perses-dev/core';
+import { PanelProps, LegendSpecOptions, OptionsEditorProps } from '@perses-dev/plugin-system';
 
 export type LogsTableProps = PanelProps<LogsTableOptions, LogsQueryData>;
 
@@ -25,6 +25,19 @@ export interface LogsTableOptions {
   thresholds?: ThresholdOptions;
   allowWrap?: boolean;
   enableDetails?: boolean;
+  pagination?: boolean;
+  showLogLine?: boolean;
+  transforms?: Transform[];
+}
+
+export interface LogsTableOptionsObsolete {
+  legend?: LegendSpecOptions;
+  thresholds?: ThresholdOptions;
+  allowWrap?: boolean;
+  enableDetails?: boolean;
   showTime?: boolean;
   showAll?: boolean;
+  transforms?: Transform[];
 }
+
+export type LogsTableSettingsEditorProps = OptionsEditorProps<LogsTableOptions>;
